+++ services/accountService.js
@@ -2,6 +2,7 @@ import UserState from '../models/UserState.js';
 import AccessGrant from '../models/AccessGrant.js';
 import AccountToken from '../models/AccountToken.js';
 import AiUsageDaily from '../models/AiUsageDaily.js';
+import AiCacheEntry from '../models/AiCacheEntry.js';
 import ConversationRoom from '../models/ConversationRoom.js';
 import AiContentReport from '../models/AiContentReport.js';
 import ConversationReport from '../models/ConversationReport.js';
@@ -19,6 +20,7 @@ export async function deleteUserAccount(user) {
     UserState.deleteOne({ user: user._id }),
     AccountToken.deleteMany({ user: user._id }),
     AiUsageDaily.deleteMany({ user: user._id }),
+    AiCacheEntry.deleteMany({ user: user._id }),
     ConversationRoom.deleteMany({ hostUser: user._id }),
     AiContentReport.deleteMany({ user: user._id }),
     ConversationReport.deleteMany({ hostUser: user._id }),